feat: ajutes nos estilos e organizacao
