--- frontend/src/components/Main.jsx
@@ -1,22 +1,0 @@
-import { Hero } from "./Hero";
-import { Carousel } from './Carousel';
-import { Values } from './Values';
-import { Avaliation } from "./Avaliation";
-import { ContactForm } from "./ContactForm";
-import { Footer } from './Footer';
-function Main() {
-
-  return (
-    <>
-      <Hero />  
-      <Carousel />
-      <Values />
-      <Avaliation />
-      <ContactForm  />
-      <Footer />
-
-    </>
-  );
-}
-
-export { Main };
