+++ frontend/src/utils/api.js
@@ -1,0 +1,10 @@
+import axios from 'axios';
+
+const api = axios.create({
+  baseURL: 'http://localhost:8080',
+  headers: {
+    'Content-Type': 'application/json',
+  },
+});
+
+export default api;
