+++ package.json
@@ -1,0 +1,5 @@
+{
+  "dependencies": {
+    "axios": "^1.13.2"
+  }
+}
